requirements.txt
pymongo
openai
python-dotenv